Feature/ncrm 513

See merge request ecosystem/users.service!9

+++ service/rolesService.js
@@ -15,14 +15,14 @@ module.exports = {
   /**
    * Get role by id
    * @param roleId
-   * @param withRights
+   * @param appends
    * @param Models
    * @return {Promise<*>}
    */
-  getRole: async(roleId, withRights, {Models}) => {
+  getRole: async(roleId, appends = [], {Models}) => {
     const options = {};
 
-    if (withRights) {
+    if (appends.includes('rights')) {
       options.include = [{
         model: Models.Right,
         as: 'rights',
@@ -45,10 +45,10 @@ module.exports = {
    * @param limit
    * @param offset
    * @param order
-   * @param withRights
+   * @param appends
    * @param Models
    */
-  searchRoles: async({where, limit, offset, order}, withRights, {Models}) => {
+  searchRoles: async({where, limit, offset, order}, appends = [], {Models}) => {
     const options = {
       limit,
       offset,
@@ -56,7 +56,7 @@ module.exports = {
       order
     };
 
-    if (withRights) {
+    if (appends.includes('rights')) {
       options.include = [{
         model: Models.Right,
         as: 'rights',
@@ -198,8 +198,9 @@ module.exports = {
   setRoleRights: async(roleId, rightsIds, {Models}) =>
     await Models.sequelize.transaction(async(transaction) => {
       await Models.RoleRight.destroy({
-        where: {roleId: {[Models.Sequelize.Op.eq]: roleId}}
-      }, {transaction});
+        where: {roleId: {[Models.Sequelize.Op.eq]: roleId}},
+        transaction
+      });
 
       const rolesRights = await Models.RoleRight.bulkCreate(rightsIds.map((rightId) => {
         return {roleId, rightId};
@@ -253,15 +254,24 @@ module.exports = {
    * @return {Promise<*>}
    */
   setUserRoles: async(userId, rolesIds, {Models}) => {
-    await Models.UserRole.destroy({
-      where: {userId: {[Models.Sequelize.Op.eq]: userId}}
+    const user = await Models.User.findByPk(userId);
+
+    if (!user) {
+      throw new Error(`user with id ${userId} is not exist`);
+    }
+
+    return await Models.sequelize.transaction(async(transaction) => {
+      await Models.UserRole.destroy({
+        where: {userId: {[Models.Sequelize.Op.eq]: userId}},
+        transaction
+      });
+
+      const userRoles = await Models.UserRole.bulkCreate(rolesIds.map((roleId) => {
+        return {userId, roleId};
+      }), {returning: true, transaction});
+
+      return userRoles.map((userRole) => userRole.get({plain: true}));
     });
-
-    const userRoles = await Models.UserRole.bulkCreate(rolesIds.map((roleId) => {
-      return {userId, roleId};
-    }), {returning: true});
-
-    return userRoles.map((userRole) => userRole.get({plain: true}));
   },
   /**
    * Add role for user
@@ -271,6 +281,12 @@ module.exports = {
    * @return {Promise<*>}
    */
   addUserRole: async(userId, roleId, {Models}) => {
+    const user = await Models.User.findByPk(userId);
+
+    if (!user) {
+      throw new Error(`user with id ${userId} is not exist`);
+    }
+
     const userRole = await Models.UserRole.create({userId, roleId});
 
     return userRole.get({plain: true});
@@ -318,5 +334,22 @@ module.exports = {
     }
 
     return role.rights.map((item) => item.get());
+  },
+  getUserRoles: async(userId, {Models}) => {
+    const user = await Models.User.findByPk(userId, {
+      include: [{
+        model: Models.Role,
+        as: 'roles',
+        through: {
+          attributes: []
+        }
+      }]
+    });
+
+    if (!user) {
+      throw new Error(`user with id ${userId} is missing`);
+    }
+
+    return user.roles.map((item) => item.get({plain: true}));
   }
 };
